Refactoring and code cleanup
